pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tweenforge"
version = "0.1.0"
description = "AI-powered inbetween frame generator for Clip Studio Paint animators"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "pillow>=10.0.0",
    "numpy>=1.26.0",
    "torch>=2.1.0",
    "torchvision>=0.16.0",
    "pydantic>=2.5.0",
    "click>=8.1.0",
    "httpx>=0.27.0",
    "python-multipart>=0.0.9",
]

[project.optional-dependencies]
cloud = [
    "boto3>=1.34.0",
    "gunicorn>=21.2.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.3.0",
]

[project.scripts]
tweenforge = "tweenforge.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/tweenforge"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
